refactor(cms): move location cover field into edit sheet
Remove the separate LocationCoverEditor dialog and "Edit cover image"
row action. Replace with a LocationCoverField inline component rendered
via a new extraContent slot on EditSheet. Also stores editLocationId
instead of a LocationRow snapshot so the cover thumbnail stays current
after linkMedia/clearCover mutations without closing the sheet.
