+++ frontend/src/types/models/location.types.ts
@@ -67,6 +67,7 @@ export type LocationRow = {
     isOwnedByViernulvier: boolean | null;
     uitdatabankId: string | null;
     address: string;
+    coverImageUrl: string | null;
     descriptionNl: string | null;
     descriptionEn: string | null;
     historyNl: string | null;
